+++ Core.cs
@@ -23,13 +23,13 @@ internal class Core
     static Entity _localCharacter = Entity.Null;
     static Entity _localUser = Entity.Null;
     public static Entity LocalCharacter =>
-        _localCharacter != Entity.Null
+        _localCharacter.Exists()
         ? _localCharacter
         : (ConsoleShared.TryGetLocalCharacterInCurrentWorld(out _localCharacter, _client)
         ? _localCharacter
         : Entity.Null);
     public static Entity LocalUser =>
-        _localUser != Entity.Null
+        _localUser.Exists()
         ? _localUser
         : (ConsoleShared.TryGetLocalUserInCurrentWorld(out _localUser, _client)
         ? _localUser